requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dash==2.10.2
dash-bootstrap-components==1.4.1
dash-bootstrap-templates==1.0.8
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
docopt==0.6.2
Flask==2.2.5
h5py==3.8.0
idna==3.4
infinity==1.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
llvmlite==0.40.1rc1
MarkupSafe==2.1.3
numba==0.57.0
numpy==1.24.3
packaging==23.1
pipreqs==0.4.13
plotly==5.15.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sortedcontainers==2.4.0
tenacity==8.2.2
threadpoolctl==3.1.0
traces==0.6.0
tslearn==0.5.3.2
urllib3==2.0.3
Werkzeug==2.2.3
yarg==0.1.9
